third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/drawing-display-none-fails.html

<!DOCTYPE html>
<title>Drawing an element with display:none should fail</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<canvas id="canvas_2d" layoutsubtree width="100" height="100" style="background: green;">
  <div id="canvas_2d_div" style="width: 100px; height: 100px; background: red;"></div>
</canvas>

<canvas id="canvas_3d" layoutsubtree width="100" height="100" style="background: green;">
  <div id="canvas_3d_div" style="width: 100px; height: 100px; background: red;"></div>
</canvas>

<script>
  test(function() {
    canvas_2d_div.getBoundingClientRect();
    canvas_2d_div.style.display = 'none';
    const context_2d = canvas_2d.getContext('2d');
    assert_throws_js(
      TypeError,
      () => context_2d.drawElementImage(canvas_2d_div, 0, 0),
      'drawElementImage cannot draw display:none elements.'
    );
  }, 'drawElementImage with display:none element');

  test(function() {
    canvas_3d_div.getBoundingClientRect();
    canvas_3d_div.style.display = 'none';
    const context_gl = canvas_3d.getContext('webgl2');
    const tex = context_gl.createTexture();
    context_gl.bindTexture(context_gl.TEXTURE_2D, tex);
    const level = 0;
    const internalformat = context_gl.RGBA;
    const format = context_gl.RGBA;
    const type = context_gl.UNSIGNED_BYTE;
    assert_throws_js(
      TypeError,
        () => context_gl.texElementImage2D(
            context_gl.TEXTURE_2D, level, internalformat, format, type,
            canvas_3d_div),
      'texElementImage2D cannot draw display:none elements.'
    );
  }, 'texElementImage2D with display:none element');
</script>
